Move validation rules into config dir
